requirements.txt
sentry-sdk
motor
pydantic
httpx
msal
pyjwt>=2.8.0
aiocache
apscheduler
motor
pyopenssl
cryptography==38.0.4
pyOpenSSL==22.1.0
openpyxl
twilio
fastapi
azure-identity
azure-keyvault-secrets